Fixes #155: Fixed error and repositioned login and help buttons

+++ Estates Audit/Controller/LoginViewController.m
@@ -81,7 +81,7 @@
             
             NSHTTPURLResponse *httpResponse = (NSHTTPURLResponse *) response;
         
-            int responseStatusCode = [httpResponse statusCode];
+            NSInteger responseStatusCode = [httpResponse statusCode];
             
             // Check if invalid
             if(responseStatusCode == 403){
